FolderをHomeからコンポーネント化
