+++ app/src/Home.js
@@ -1,9 +1,9 @@
 import React,{useState,useEffect} from "react";
 import axios from "axios"
-import {Box,Button,Center,Heading} from "@chakra-ui/react"
-import {ReactSortable} from "react-sortablejs"
+import {Box,Button,Heading} from "@chakra-ui/react"
 import Tree from "./Tree"
 import Set from "./Set"
+import Folder from "./Folder";
 
 export default function App(user_id) {
   const [group,setGroup] = useState([])
@@ -20,17 +20,9 @@ export default function App(user_id) {
       else if (Array.isArray(info[0].children)){setPage("Tree")}
     }
   },[info])
-
-  function ClickFile(folder_id){
-    axios.get(`http://127.0.0.1:8010/contents/${folder_id}`)
-      .then((res) => {
-        setInfo(res.data.contents)
-      })
-      .catch((err) => {console.log("contentsAPI err:",err)})
-  }
-
   
   if (page == "Home"){
+    console.log(info)
     return (
       <Box>
         <Box
@@ -44,23 +36,7 @@ export default function App(user_id) {
         >
           <Heading float="left" color="white">Bookmark</Heading>
         </Box>
-        <Box align='center' py="90px" maxWidth="1500px" margin="auto">
-        <ReactSortable
-        list={group}
-        setList={setGroup}
-        animation={150}
-        onChange={(order, sortable, evt) => {}}
-        onEnd={evt => {}}
-        align='center' justify='center'
-        >
-          {group.map((g) => 
-          <Box display="inline-block" padding="10px">
-            <Center onClick={() => ClickFile(g.id)} bg='#CBD5E0' width='200px' height='100px'　borderRadius='10px' _hover={{bg:'#A0AEC0',cursor:'pointer'}}>
-              {g.title}
-            </Center>
-          </Box>)}
-        </ReactSortable>
-        </Box>
+        {Folder(group,setGroup,info,setInfo)}
       </Box>
     )
   }
